+++ CHANGELOG.md
@@ -7,11 +7,13 @@ Bugs fixed:
 
 New Features:
 
+- (api) Add PlayerDeathCallback
+- (api) Add PlayerPreDeathCallback
 - (module-teleport) Add death support to /back
 
 Improvements:
 
-- (api) Add PlayerDeathCallback (+ updated onDeath mixins to use this callback)
+- None
 
 
 ## Essentials 0.2.3